Changed indexes when referencing insertedIds in Mutations
I love the tutorial, but was confused why the example mutation query wasn't running off-the-bat.

This error was popping  up:

```
TypeError: Cannot read property '_id' of null
```

A console log of the query in the resolver showed that the id was the only item in the array `insertedIds` array.  Changing the index to the first item fixed it.  Maybe related to the version of MongoDB released at the time the tutorial was published.